+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for http-interchange
 
+## 0.3.1.0 -- 2023-08-16
+
+* Add these to `Http.Headers`: cons, snoc, lookupHost, lookupAccept,
+  lookupDate, lacksContentLengthAndTransferEncoding,
+  snocContentLength.
+
 ## 0.3.0.0 -- 2023-08-15
 
 * Add `Http.Headers.fromList`. In examples, this makes it easier